+++ Sandbox/Examples/Testing/GUITesting.h
@@ -4,7 +4,7 @@ class GUITestingLayer :public Layer {
 public:
     GUITestingLayer();
     void OnCreate()override;
-    void OnUpdate(double deltatime)override;
+    void OnUpdate(float deltatime)override;
     void OnGUI()override;
     void OnDestroy()override;
